Closes #451 (gabrielfalcao/lettuce/issues/451) (#518)
New hook for outline added. It works before_each outline and after_each outline. It is required for PyCharm lettuce scenario outline support, and it is useful anyway.

Old hook is left unchanged to keep old API, but it can't be used because it is not called before and after each outline, and in case of exception and fastfail it is called after after_scenario.